Added EditCampus and EditStudent component. Only worked on EditCampus so far. It currently allows a user to update the campus information. When "Save Changes" is clicked, it takes the user back to the campus view page.
Also added "add student" functionality to the page. Only students not assigned to a campus can be added to the current campus. Students currently enrolled in the campus are shown below. If there are not unassigned students, then the form is removed and replaced with a message indicating so. The same is true for assigned students.

If at any point when saving changes or adding a student an error occurs, then this error will be printed. This can happen if any of the input boxes are empty.